Fix recursive distinct, array return, intrinsic types, keep, shift UB (BUG-314/315/316/317/318)
BUG-314: type_unwrap_distinct in recursive struct/union check — distinct
  typedef wrapping Self no longer bypasses self-reference detection
BUG-315: type_unwrap_distinct on return type before TYPE_ARRAY check —
  distinct typedef u8[10] Buffer correctly rejected as array return
BUG-316: force_type_arg for @bitcast/@truncate/@saturate — named types
  (distinct typedefs) now accepted as first argument, not just @cast
BUG-317: keep validation checks mangled key (module_name) when raw
  lookup fails — imported globals no longer falsely rejected
BUG-318: constant folder shift uses (uint64_t) cast — prevents signed
  overflow UB in the compiler itself on large left-shift operands

+++ ast.h
@@ -519,7 +519,8 @@ static inline int64_t eval_const_expr(Node *n) {
             return l % r;
         case TOK_LSHIFT:
             if (r < 0 || r >= 63) return CONST_EVAL_FAIL;
-            return l << r;
+            /* BUG-318: cast to unsigned to avoid signed overflow UB */
+            return (int64_t)((uint64_t)l << r);
         case TOK_RSHIFT:
             if (r < 0 || r >= 63) return CONST_EVAL_FAIL;
             return l >> r;
